requirements.txt
kafka-python==2.0.2
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
logging==0.4.9.6